+++ include/Engine/Renderer/Mesh/SkeletalMesh.h
@@ -2,7 +2,7 @@
 //	Created by MarcasRealAccount on 30. Oct. 2020
 //	
 
-#include "Engine/Renderer/Mesh.h"
+#include "Engine/Renderer/Mesh/Mesh.h"
 
 namespace gp1 {
 
@@ -20,17 +20,6 @@ namespace gp1 {
 
 	public:
 		std::vector<SkeletalMeshVertex> m_Vertices;	// This mesh's vertices.
-	};
-
-	struct OpenGLSkeletalMeshData : public OpenGLMeshData {
-	public:
-		OpenGLSkeletalMeshData(Mesh* mesh);
-
-	private:
-		bool HasVertices() override;
-		uint32_t GetCustomDataSize() override;
-		void InitCustomGLData() override;
-		void ClearCustomData() override;
 	};
 
 } // namespace gp1
