+++ include/Engine/Renderer/Mesh/StaticMesh.h
@@ -2,7 +2,7 @@
 //	Created by MarcasRealAccount on 30. Oct. 2020
 //	
 
-#include "Engine/Renderer/Mesh.h"
+#include "Engine/Renderer/Mesh/Mesh.h"
 
 namespace gp1 {
 
@@ -18,17 +18,6 @@ namespace gp1 {
 
 	public:
 		std::vector<StaticMeshVertex> m_Vertices;	// This mesh's vertices.
-	};
-
-	struct OpenGLStaticMeshData : public OpenGLMeshData {
-	public:
-		OpenGLStaticMeshData(Mesh* mesh);
-
-	private:
-		bool HasVertices() override;
-		uint32_t GetCustomDataSize() override;
-		void InitCustomGLData() override;
-		void ClearCustomData() override;
 	};
 
 } // namespace gp1
