+++ include/Engine/Renderer/Shader/Material.h
@@ -4,12 +4,9 @@
 
 #pragma once
 #include "Engine/Renderer/Renderer.h"
-#include "Engine/Renderer/Shader/Uniform.h"
 
 #include <unordered_map>
 #include <any>
-
-#include <glad/glad.h>
 
 namespace gp1 {
 
@@ -174,43 +171,18 @@ namespace gp1 {
 
 		virtual RendererType GetRendererType() const = 0;
 
+		friend Renderer;
+
 	protected:
-		Material* GetMaterial() const;
+		template <typename T>
+		T* GetMaterial() const {
+			return reinterpret_cast<T*>(this->m_Material);
+		}
 
 		const std::unordered_map<std::string, std::any>& GetUniforms() const;
 
 	private:
 		Material* m_Material;
-	};
-
-	class OpenGLMaterialData : public MaterialData {
-	public:
-		OpenGLMaterialData(Material* material);
-
-		virtual void CleanUp() override;
-
-		virtual RendererType GetRendererType() const override;
-
-		// Get the cull face.
-		GLenum GetCullFace() const;
-		// Get the src blend function.
-		GLenum GetSrcBlendFunc() const;
-		// Get the dst blend function.
-		GLenum GetDstBlendFunc() const;
-		// Get the face to use the polygon mode on.
-		GLenum GetPolygonModeFace() const;
-		// Get the polygon mode.
-		GLenum GetPolygonMode() const;
-
-		// Set all uniforms.
-		void SetAllUniforms();
-
-		friend OpenGLRenderer;
-
-	private:
-		static GLenum GetGLCullFace(TriangleFace face);
-		static GLenum GetGLBlendFunc(BlendFunc blendFunc);
-		static GLenum GetGLPolygonMode(PolygonMode polygonMode);
 	};
 
 } // namespace gp1
