+++ src/Engine/Renderer/Apis/OpenGL/Mesh/OpenGLMeshData.cpp
@@ -1,59 +1,10 @@
 //	
-//	Created by MarcasRealAccount on 29. Oct. 2020
+//	Created by MarcasRealAccount on 31. Oct. 2020
 //	
 
-#include "Engine/Renderer/Mesh.h"
+#include "Engine/Renderer/Apis/OpenGL/Mesh/OpenGLMeshData.h"
 
 namespace gp1 {
-
-	Mesh::~Mesh() {
-		if (this->m_MeshData) {
-			this->m_MeshData->CleanUp();
-			delete this->m_MeshData;
-		}
-	}
-
-	void Mesh::MarkDirty() {
-		this->m_Dirty = this->m_Editable;
-	}
-
-	void Mesh::ClearDirty() {
-		this->m_Dirty = false;
-	}
-
-	bool Mesh::IsDirty() {
-		return this->m_Dirty;
-	}
-
-	bool Mesh::IsEditable() {
-		return this->m_Editable;
-	}
-
-	bool Mesh::IsDynamic() {
-		return this->m_IsDynamic;
-	}
-
-	MeshData* Mesh::GetMeshData(Renderer* renderer) {
-		if (!this->m_MeshData || !renderer->IsMeshDataUsable(this->m_MeshData)) {
-			if (this->m_MeshData) {
-				this->m_MeshData->CleanUp();
-				delete this->m_MeshData;
-			}
-			this->m_MeshData = CreateCustomMeshData(renderer);
-		}
-		return this->m_MeshData;
-	}
-
-	MeshData::MeshData(Mesh* mesh)
-		: m_Mesh(mesh) {}
-
-	MeshData::~MeshData() {
-		this->m_Mesh->m_MeshData = nullptr;
-	}
-
-	Mesh* MeshData::GetMesh() const {
-		return this->m_Mesh;
-	}
 
 	OpenGLMeshData::OpenGLMeshData(Mesh* mesh)
 		: MeshData(mesh) {}
@@ -67,7 +18,7 @@ namespace gp1 {
 	}
 
 	uint32_t OpenGLMeshData::GetVAO() {
-		if (GetMesh()->IsDirty())
+		if (GetMesh<Mesh>()->IsDirty())
 			InitGLData();
 		return this->m_VAO;
 	}
@@ -79,7 +30,7 @@ namespace gp1 {
 		if (!HasVertices())
 			return;
 
-		this->m_HasIndices = GetMesh()->m_Indices.size() > 0;
+		this->m_HasIndices = GetMesh<Mesh>()->m_Indices.size() > 0;
 
 		glGenVertexArrays(1, &this->m_VAO);
 		glBindVertexArray(this->m_VAO);
@@ -90,19 +41,19 @@ namespace gp1 {
 
 		if (this->m_HasIndices) {
 			BindNextVBO(GL_ELEMENT_ARRAY_BUFFER);
-			glBufferData(GL_ELEMENT_ARRAY_BUFFER, GetMesh()->m_Indices.size() * sizeof(uint32_t), GetMesh()->m_Indices.data(), GL_STATIC_DRAW);
-			this->m_BufferSize = static_cast<uint32_t>(GetMesh()->m_Indices.size());
+			glBufferData(GL_ELEMENT_ARRAY_BUFFER, GetMesh<Mesh>()->m_Indices.size() * sizeof(uint32_t), GetMesh<Mesh>()->m_Indices.data(), GL_STATIC_DRAW);
+			this->m_BufferSize = static_cast<uint32_t>(GetMesh<Mesh>()->m_Indices.size());
 		} else {
 			this->m_BufferSize = GetCustomDataSize();
 		}
 
 		glBindVertexArray(0);
 
-		if (!GetMesh()->IsEditable() || !GetMesh()->IsDynamic()) {
+		if (!GetMesh<Mesh>()->IsEditable() || !GetMesh<Mesh>()->IsDynamic()) {
 			ClearCustomData();
-			GetMesh()->m_Indices.clear();
+			GetMesh<Mesh>()->m_Indices.clear();
 		}
-		GetMesh()->ClearDirty();
+		GetMesh<Mesh>()->ClearDirty();
 	}
 
 	void OpenGLMeshData::CleanUp() {
