+++ src/Engine/Renderer/Apis/OpenGL/Mesh/OpenGLStaticVoxelMeshData.cpp
@@ -1,30 +1,26 @@
 //	
-//	Created by MarcasRealAccount on 30. Oct. 2020
+//	Created by MarcasRealAccount on 31. Oct. 2020
 //	
 
-#include "Engine/Renderer/StaticVoxelMesh.h"
+#include "Engine/Renderer/Apis/OpenGL/Mesh/OpenGLStaticVoxelMeshData.h"
 
 namespace gp1 {
-
-	MeshData* StaticVoxelMesh::CreateCustomMeshData(Renderer* renderer) {
-		return renderer->CreateStaticVoxelMeshData(this);
-	}
 
 	OpenGLStaticVoxelMeshData::OpenGLStaticVoxelMeshData(Mesh* mesh)
 		: OpenGLMeshData(mesh) {}
 
 	bool OpenGLStaticVoxelMeshData::HasVertices() {
-		return reinterpret_cast<StaticVoxelMesh*>(GetMesh())->m_Vertices.size() > 0;
+		return GetMesh<StaticVoxelMesh>()->m_Vertices.size() > 0;
 	}
 
 	uint32_t OpenGLStaticVoxelMeshData::GetCustomDataSize() {
-		return static_cast<uint32_t>(reinterpret_cast<StaticVoxelMesh*>(GetMesh())->m_Vertices.size());
+		return static_cast<uint32_t>(GetMesh<StaticVoxelMesh>()->m_Vertices.size());
 	}
 
 	void OpenGLStaticVoxelMeshData::InitCustomGLData() {
 		CreateVBOs(this->m_NumVBOs + 1);
 		BindNextVBO(GL_ARRAY_BUFFER);
-		glBufferData(GL_ARRAY_BUFFER, reinterpret_cast<StaticVoxelMesh*>(GetMesh())->m_Vertices.size() * sizeof(StaticVoxelMeshVertex), reinterpret_cast<StaticVoxelMesh*>(GetMesh())->m_Vertices.data(), GL_STATIC_DRAW);
+		glBufferData(GL_ARRAY_BUFFER, GetMesh<StaticVoxelMesh>()->m_Vertices.size() * sizeof(StaticVoxelMeshVertex), GetMesh<StaticVoxelMesh>()->m_Vertices.data(), GL_STATIC_DRAW);
 		CreateVertexAttribArrays(4);
 		SetVertexAttribPointer(static_cast<uint32_t>(VertexAttribIndex::POSITION), 3, GL_FLOAT, false, sizeof(StaticVoxelMeshVertex), offsetof(StaticVoxelMeshVertex, position));
 		SetVertexAttribPointer(static_cast<uint32_t>(VertexAttribIndex::NORMAL), 3, GL_FLOAT, false, sizeof(StaticVoxelMeshVertex), offsetof(StaticVoxelMeshVertex, normal));
@@ -34,7 +30,7 @@ namespace gp1 {
 	}
 
 	void OpenGLStaticVoxelMeshData::ClearCustomData() {
-		reinterpret_cast<StaticVoxelMesh*>(GetMesh())->m_Vertices.clear();
+		GetMesh<StaticVoxelMesh>()->m_Vertices.clear();
 	}
 
 } // namespace gp1
