+++ src/Engine/Renderer/Apis/Vulkan/Mesh/VulkanSkeletalMeshData.cpp
@@ -1,0 +1,12 @@
+//	
+//	Created by MarcasRealAccount on 31. Oct. 2020
+//	
+
+#include "Engine/Renderer/Apis/Vulkan/Mesh/VulkanSkeletalMeshData.h"
+
+namespace gp1 {
+
+	VulkanSkeletalMeshData::VulkanSkeletalMeshData(Mesh* mesh)
+		: VulkanMeshData(mesh) {}
+
+} // namespace gp1
