+++ src/Engine/Renderer/Apis/Vulkan/Mesh/VulkanStaticMeshData.cpp
@@ -1,0 +1,12 @@
+//	
+//	Created by MarcasRealAccount on 31. Oct. 2020
+//	
+
+#include "Engine/Renderer/Apis/Vulkan/Mesh/VulkanStaticMeshData.h"
+
+namespace gp1 {
+
+	VulkanStaticMeshData::VulkanStaticMeshData(Mesh* mesh)
+		: VulkanMeshData(mesh) {}
+
+} // namespace gp1
