+++ src/Engine/Renderer/Apis/Vulkan/Mesh/VulkanStaticVoxelMeshData.cpp
@@ -1,0 +1,12 @@
+//	
+//	Created by MarcasRealAccount on 31. Oct. 2020
+//	
+
+#include "Engine/Renderer/Apis/Vulkan/Mesh/VulkanStaticVoxelMeshData.h"
+
+namespace gp1 {
+
+	VulkanStaticVoxelMeshData::VulkanStaticVoxelMeshData(Mesh* mesh)
+		: VulkanMeshData(mesh) {}
+
+} // namespace gp1
